requirements.txt
telethon
python-dotenv
beautifulsoup4
pandas
numpy
matplotlib
seaborn
emoji
sqlalchemy
psycopg2-binary
opencv-python
torch
torchvision
tensorflow
fastapi 
uvicorn 
psycopg2
